Add support for parcel v2
