+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@riotjs/parcel-plugin-riot",
+  "name": "@riotjs/parcel-transformer-riot",
   "version": "6.0.0",
   "description": "A parcel plugin for riot.js",
   "main": "index.js",
@@ -8,17 +8,23 @@
     "test": "npx eslint *.js && npm run test-bundle",
     "test-bundle": "cd test && npm i"
   },
+  "engines": {
+    "parcel": "2.x"
+  },
   "license": "MIT",
   "repository": {
     "type": "git",
     "url": "https://github.com/andruschka/parcel-plugin-riot"
   },
   "peerDependencies": {
-    "@riotjs/compiler": "^6.0.0",
-    "parcel-bundler": "^1.9.4"
+    "@riotjs/compiler": "^6.0.0"
   },
   "devDependencies": {
     "eslint": "^7.30.0",
     "eslint-config-riot": "^3.0.0"
+  },
+  "dependencies": {
+    "@parcel/plugin": "^2.0.0",
+    "@parcel/source-map": "^2.0.0"
   }
 }
