+++ test/.parcelrc
@@ -1,0 +1,6 @@
+{
+	"extends": "@parcel/config-default",
+	"transformers": {
+		"*.riot": ["@riotjs/parcel-transformer-riot"]
+	}
+}
