+++ test/index.html
@@ -8,6 +8,6 @@
 </head>
 <body>
     <div id="root"></div>
-    <script src="./main.js"></script>
+    <script type="module" src="./main.js"></script>
 </body>
 </html>
