+++ test/package.json
@@ -1,17 +1,16 @@
 {
   "name": "riot-parcel-test",
-  "main": "index.html",
   "private": true,
   "scripts": {
     "postinstall": "parcel build index.html --no-cache"
   },
   "dependencies": {
-    "parcel-bundler": "^1.11.0"
+    "parcel": "^2.0.0"
   },
   "devDependencies": {
-    "@riotjs/compiler": "^4.0.0",
-    "@riotjs/hot-reload": "^4.0.0-rc.1",
-    "@riotjs/parcel-plugin-riot": "file:../",
-    "riot": "^4.0.0-rc.13"
+    "@riotjs/compiler": "^6.0.0",
+    "@riotjs/hot-reload": "^6.0.0",
+    "@riotjs/parcel-transformer-riot": "file:../",
+    "riot": "^6.0.0"
   }
 }
